Refactor Slack into seperate app and allow per-channel tokens
