+++ evewspace/Jabber/models.py
@@ -26,16 +26,3 @@ class JabberAccount(models.Model):
 
     def __unicode__(self):
         return "User: %s  JID: %s" % (self.user.username, self.jid)
-
-class SlackChannel(models.Model):
-    """
-    Reference Slack channel to send messages to based on SubGroup name.
-    """
-    channel = models.CharField(max_length=50, null=False, unique=True)
-    group = models.ForeignKey(SubscriptionGroup, related_name='slack_groups', null=False, unique=True)
-    
-    
-    def __unicode__(self):
-        return "Channel: %s Group: %s" % (self.channel, self.group.name)
-
-
