+++ evewspace/Slack/alert_methods.py
@@ -1,0 +1,3 @@
+from Alerts.method_registry import registry
+from slack_method import SlackAlertMethod
+registry.register("Slack", SlackAlertMethod)
