+++ requirements.txt
@@ -1,5 +1,4 @@
 Django==1.6.1
-MySQL-python==1.2.5
 South==0.8.2
 amqp==1.0.13
 amqplib==1.0.2
@@ -19,4 +18,4 @@ sleekxmpp==1.1.11
 wsgiref==0.1.2
 networkx==1.8.1
 PyTS3==0.3
-requests=2.4.0
+requests==2.4.0
